TIMOB-10250, TIMOB-10258 and TIMOB-10384
Reviewers: David L.

Fixed following issues:
TIMOB-10250: When calling clearInterval(interval) with interval parameter
'undefined' - clearInterval() crashes.
TIMOB-10258 - itanium.UI.createImageView() is not working properly.
Creating image and adding it to the window doesn't show image
TIMOB-10384 When creating new window and trying to open it from the
current window it is not opening.

Test Cases:
- make sure build for tibb and tibbtest is not broken
- make sure when we call clearInterval fucntion with argument, which
  doesn't have number type, it is throwing an exception
- make sure createImageView is working (image property is working)
- make sure we can open new window on the top of old one