Tweak the approach to handling missing BlackBerry resources.
Rather than only fallback to the SDK templates for an empty
or missing Resources folder, just always copy the templates
into the build assets folder than copy in the project specific
files in Resources/blackberry if it exists.

This avoids the issues with missing required assets, but still
allows the project to provide its own files.

Conflicts:
	build_templates/blackberry/cli/common/blackberryndk.js

[TIMOB-14068]