Fix a bug which caused list items to disappear.
There appears to be a bug in Cascades if you try to setRoot()
on a control which was previously setRoot() on a different instance.
So to avoid this problem we wrap the row content in a container.
That container is always bound to the custom list view, so no need
to switch the root.

[TIMOB-12788]